[api] Remove api/v1 directory, move inventory_units controller to correct spot
