+++ api/app/views/spree/api/inventory_units/show.rabl
